Recorked product db, added hs

+++ hfc/service/ai_caller.go
@@ -4,5 +4,5 @@ import "github.com/eiachh/hfc/types"
 
 type AiCaller interface {
 	ParseOff([]byte) (*types.Product, error)
-	WebScrapeParse(barcode int) (*types.Product, error)
+	WebScrapeParse(barcode int64) (*types.Product, error)
 }
